[fuchsia] Force an internal image roll.
Bots are in a bad state due to GPU failures and another failure introduced that relies on a change present after 02/15.

This should help reduce the number of failures so that GPU failures can be triaged.

The internal image roller depends on the GPU tests passing, and that
may not come until after we force the roll.

Change-Id: Ic3306784836d908317233f8ff1e5a686c8adaee4
fuchsia: 121700
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4287141
Commit-Queue: Rohan Pavone <rohpavone@chromium.org>
Auto-Submit: Rohan Pavone <rohpavone@chromium.org>
Reviewed-by: David Dorwin <ddorwin@chromium.org>
Reviewed-by: David Song <wintermelons@google.com>
Cr-Commit-Position: refs/heads/main@{#1109137}
NOKEYCHECK=True
GitOrigin-RevId: 75d13c857a13561613595a263887eff956ee1138