Deprecate //:gn_visibility build target.
//:gn_visibility was used during the GYP->GN migration as a
target that targets in other files could grant visibility to
that wasn't gn_all, because for some reason we thought having
a direct dependency from gn_all to the target was too broadly
scoped.

In practice the limited visibility didn't really accomplish
anything useful, and so it's just been a kind of strange anachronism
in the build.

This CL starts the process of removing the target. We make
gn_visability an empty target and remove any references to
it in this and other repos (like pdfium and v8) that currently
contain references to it.

Once all of the references are gone, we can remove gn_visibility
itself.

Bug: 1311306
Change-Id: I7b4ee43e1653171488137d5332554795977b40e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4285777
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1109343}
NOKEYCHECK=True
GitOrigin-RevId: 0356f6f7fd30dd469330dd244ef1f8fd0c666e79