[ios] Refactor globlists to reside near associated data
With this change, we now check that files referenced by a globlist are
found in the same directory as the globlist or a subdirectory. In
general, this means that the filelists are situated alongside the data
rather than the targets that use them. This will make it less likely
for a change to add or remove a file and not have a presubmit check
catch the fact that a filelist needs updating.

There are cases where it is undesirable or inconvenient to do this,
however. It is possible to suppress these errors in those cases (eg,
for blink, we continue to glob data from an adjacent directory for the
external wpt resources).

For the components, the file lists have been moved alongside the data
in //components/test/data (with the exception of viz, since moving
this is more involved and will happen in a separate CL). Since that
directory already has permissive OWNERS, the punchouts for .filelists
could be removed from the other components directories.

For //net, a filelist has been split in order to move near its
corresponding data, with one filelist moving to //net/data and another
being added to //net/third_party/nist-pkits/.

credit: I'd originally thought of having a global PRESUBMIT to catch
all filelist updates, but dtapuska suggested keeping the filelists
near their data which appears to be a much more sustainable solution
and decreases duplication.

Bug: 1419521
Change-Id: I7cff6e2e2a0b5918f4228a6f3ac5791361833c2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4296497
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Ian Vollick <vollick@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1112034}
NOKEYCHECK=True
GitOrigin-RevId: 63109f69faee733d3ac3ed23bb8d2fc131ed93dc