Allow dylibs targeting android to be linked with rustc
Rustc attempts to link `libgcc` unconditionally on Android, but it's a
mistake to do so. We get our unwinding from `libunwind`, which we link
in explicitly ourselves, and we have no `libgcc` present. To avoid a
missing library, we point rustc to an empty `libgcc.a` file. This
allows rustc to link .so files targeting android.

To keep things simple, we simply drop an empty libgcc.a file in the
source tree, in a directory of its own, and add the directory to the
LDFLAGS when linking with rust targeting Android.

We only need to consider dylibs, since we do not build exes for
Android.

R=hans@chromium.org

Bug: 1278030
Change-Id: Ie76d72ed246a2aa0f0e3c1c7826ca402c9fa8624

Cq-Include-Trybots: luci.chromium.try:android-rust-arm32-rel,android-rust-arm64-dbg,android-rust-arm64-rel,linux-rust-x64-dbg,linux-rust-x64-rel
Change-Id: Ie76d72ed246a2aa0f0e3c1c7826ca402c9fa8624
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4321746
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1114809}
NOKEYCHECK=True
GitOrigin-RevId: 80c8f54eb538542494536bf22957b4e2687c6a7e