Use sudo instead of su to run commands as chronos for CrOS tests
Some tests are failing with `su: not found` errors on new LKGM attempts:
https://crrev.com/c/4324200

That's likely due to https://crrev.com/c/4294849, which removed
su from CrOS images. It seems "sudo" is the preferred method
for running commands as another user. So this switches to that.

Bug: 1421441, 1423191
Change-Id: I0cd9246feb9cc2fa3e40b70e8d29d4520954cc82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4326934
Reviewed-by: Sean Kau <skau@chromium.org>
Reviewed-by: Sven Zheng <svenzheng@chromium.org>
Commit-Queue: Sean Kau <skau@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1116014}
NOKEYCHECK=True
GitOrigin-RevId: 70d542c7a88991c08ed5762ddc410735c31d8f4e