[web_engine] Statically offer root-ssl-certificates to web_instances
Previously, the "root-ssl-certificates" directory had to be offered to
the WebInstanceHost component so it could be dynamically offered to a
`web_instance`, but this was not visible or enforced in the CML shards.
Instead, statically offer the directory to "#web_instances" from where
it can be offered to individual `web_instance`s as appropriate.

Bug: 1421351
Change-Id: Ia31a4ccef102605a3cc220d521d0f6ba66083308
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4351230
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: David Dorwin <ddorwin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120068}
NOKEYCHECK=True
GitOrigin-RevId: 0d15404d111edbb1d1ac7d636a7d95183cf64022