[fuchsia] Determine device target state before proceeding with reboot.
This CL:
- defines TargetState
- defines BootMode
- determines the current target state with get_target_state
- Creates boot_device, which attempts to boot the device into the
given BootMode. If the device is already in that mode, this is a no-op.

This command attempts to boot using `ffx` first, and then, failing
that, fallback to SSH for rebooting (this forces the device into
regular mode first, and then into the desired mode).

Change-Id: Ia8ce0d22dde15a85371843d656da1bc3ca134f5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4126135
Reviewed-by: Zijie He <zijiehe@google.com>
Commit-Queue: Rohan Pavone <rohpavone@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1125429}
NOKEYCHECK=True
GitOrigin-RevId: e370b912ae77d87ec1ff7ded76e4ed8da799a246