Get bindgen to run successfully on windows
On windows we pass MSVC style flags, and we need to tell libclang to
use the cl driver mode explicitly, it does not do it by default just
because it was built on/for windows.

So we add --driver-mode=cl, which then makes other clang-specific
flags fail to parse, so we work around that through a combination of
escaping the flags we want with `/clang:` prefixes, and passing
`-Wno-unused-argument` since libclang adds some clang flags
unconditionally that it then fails to parse.

Bug: 1271215
Change-Id: I119c3a4ce76f5fc43e725473870e184d2081a9e0
Cq-Include-Trybots: luci.chromium.try:android-rust-arm32-rel,android-rust-arm64-dbg,linux-rust-x64-dbg,linux-rust-x64-rel,android-rust-arm64-rel
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4382604
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1125501}
NOKEYCHECK=True
GitOrigin-RevId: 033ac01324a50a899891ba67c6b9b2c1964e7f23