Add support for Lacros Amd64 for PGO
Lacros-amd64 builder is ready at go/pgobots.

This change introduces the .pgo.txt file containing the profile
generated for lacros-amd64, and changes surrounding the file to
consume it as part of the compile process. The changes are
similar to crrev/c/4387796, but specific to amd64.

Bug: 1417071
Change-Id: I6c662029e315b50ed610de7868a9fdac36064580
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4395564
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Sven Zheng <svenzheng@chromium.org>
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1125627}
NOKEYCHECK=True
GitOrigin-RevId: 0d2216f69807383886918bc4a558d6545537cd00