Set up build config for Windows .exe files
The CRT needs to be statically linked in to avoid needing to ship a
dll file from Visual Studio along with the exe. For build scripts the
dll is not found, as the exe files are placed in obj/path/to/target/
but the dll is up in the root_out_dir. Statically linking avoids the
use of this DLL.

Add "rust_bin" to the set of BUILDCONFIG targets so that rust-linked
exes will get the same ldflags, metadata, etc as C++-linked exes do,
and that they are able to link against the same libc++ should they
depend on any C++ code. This was blocked on moving the build of
bindgen out of the Chromium GN build, which has been done.

Change-Id: I22391c30826e4fff23db307805d66636cd0bc2f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4402338
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1129408}
NOKEYCHECK=True
GitOrigin-RevId: 01e27ca5bc6ff5971ad3c65e95266f050334a74d