[rust] Conditionalize prebuilt stdlib targets.
In some cases (e.g. Android) we don't distribute prebuilt Rust standard
libraries, so we have no choice but to build the stdlib from scratch.
This was already the case, and Rust targets linked fine on such
platforms. But, in the case of builds which targeted all gn targets, we
would have tried to run the build steps pertaining to the prebuilt Rust
stdlib which didn't exist, and such builds failed.

This CL ensures that all such build steps are behind conditions to
ensure that they are never triggered in cases where such a prebuilt
stdlib doesn't exist.

For ease of review, this CL does not restructure the file at all. The
next CL in the series reshuffles the lines to reduce the excessive
number of conditionals added by this CL.

This should have no functional effect except to disable some broken gn
targets on some builds.

Bug: 1292038
Change-Id: I7dc9c61e70e952d325d283c4020053f42d268c93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4427985
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1131325}
NOKEYCHECK=True
GitOrigin-RevId: 1a7b44516a80e91276600b0f24780212adc43e67