[rust] Cease duplicate inclusion of Rust stdlib
Back when we had a prebuilt stdlib, we were unable to feed it into gn as
a linkable file for downstream targets. For that reason we included the
filenames in ldflags.

Now we're building the Rust stdlib, they're just normal first-class gn
targets and will be propagated through dependencies in the correct way.

This appears to solve a couple of problems showing up on bots:
* Some build configurations are complaining due to duplicate symbols,
  possibly from this double-inclusion
* Some build configurations are complaining because targets include
  the Rust standard library but not the remap_alloc.cc shims. This is
  because remap_alloc, as a standard gn target, is subject to all the
  nuanced rules about propagation between different shared object
  targets, whereas ldflags aren't.

This CL simplifies things so that the Rust standard library .rlibs and
remap_alloc.cc always go along together.

Bug: 1434643, 1292038
Change-Id: Ia8e0857c382c1100dab6171d769844a3788e4774
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4447478
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1132629}
NOKEYCHECK=True
GitOrigin-RevId: 724458e64265922b3f554bd51296d20f3f76b57d