[Fuchsia] Set daemon.autostart=false
This change controls the lifespan of the ffx daemon by explicitly
calling stop at the beginning of the test, restart it with the correct
configurations and stop at the end of the test.
So that the daemon.autostart can be set to false and any unexpected
behaviors of bringing up the daemon will be blocked.

Meanwhile, it adds a start_ffx_daemon and stop_ffx_daemon to detect the
state of the daemon after starting or stopping it. Ideally the
_run_repair_command can be removed, but I won't go too far in this CL
and it will be handled later.

Two bugs have been added into the comments so that after both of them
being fixed within fuchsia-sdk, we can also remove the daemon status
check from our side.

Bug: fuchsia:92296, fuchsia:125873
Change-Id: Ifae8d76642f450c47f8d2e53ea3869d70e7f7c53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4449826
Reviewed-by: Rohan Pavone <rohpavone@chromium.org>
Commit-Queue: Zijie He <zijiehe@google.com>
Cr-Commit-Position: refs/heads/main@{#1135601}
NOKEYCHECK=True
GitOrigin-RevId: 92b247abe4d2219f4b11de089f6fd5aca94c63e3